+++ R/tryCatchLog.R
@@ -29,6 +29,9 @@
 #' Conditions are also logged including the function call stack
 #' with file names and line numbers (if available).
 #'
+#' By default the maximum number of source code rows that are printed per call in the full stack trace
+#' is 10. You can change this via the option \code{tryCatchLog.max.lines.per.call} (see example).
+#'
 #' @param expr                  R expression to be evaluated
 #' @param ...                   condition handler functions (as in \code{\link{tryCatch}}).
 #'                              The following condition names are mainly used in R:
@@ -66,6 +69,9 @@
 #'                     call stacks the message text will be output without call stacks.
 #'                     The default value can be changed globally by setting the option \code{tryCatchLog.include.compact.call.stack}.
 #'                     The compact call stack can always be found via \code{\link{last.tryCatchLog.result}}.
+#' @param logged.conditions \code{NULL}: Conditions are not logged.\cr
+#'                          \code{vector of strings}: Only conditions whose class name is contained in this vector are logged.\cr
+#'                          \code{NA}: All conditions are logged.
 #' @return                     the value of the expression passed in as parameter "expr"
 #'
 #' @details This function shall overcome some drawbacks of the standard \code{\link{tryCatch}} function.\cr
@@ -165,6 +171,8 @@
 #' options("tryCatchLog.write.error.dump.folder" = "my_log")
 #' options("tryCatchLog.write.error.dump.folder" = NULL)
 #'
+#' options(tryCatchLog.max.lines.per.call = 30)
+#'
 #' \dontrun{
 #' # Use case for "execution.context.msg" argument: Loops and parallel execution
 #' library(foreach)       # support parallel execution (requires an parallel execution plan)
@@ -183,7 +191,8 @@ tryCatchLog <- function(expr,
                         silent.warnings            = getOption("tryCatchLog.silent.warnings", FALSE),
                         silent.messages            = getOption("tryCatchLog.silent.messages", FALSE),
                         include.full.call.stack    = getOption("tryCatchLog.include.full.call.stack", TRUE),
-                        include.compact.call.stack = getOption("tryCatchLog.include.compact.call.stack", TRUE)
+                        include.compact.call.stack = getOption("tryCatchLog.include.compact.call.stack", TRUE),
+                        logged.conditions          = getOption("tryCatchLog.logged.conditions", NULL)
 ) {
 
   reset.last.tryCatchLog.result()   # TODO If an internal error is thrown in the code above the last result will be kept. Fix this?
@@ -207,6 +216,19 @@ tryCatchLog <- function(expr,
   # closure ---------------------------------------------------------------------------------------------------------
   cond.handler <- function(c) {
 
+    # Suppress logging of a (non-standard/custom) condition?
+    # NOTE: The inheritance is checked similar to the "severity" below but intentional to separate orthogonal logic
+    if (inherits(c, "condition") && !inherits(c, c("error", "warning", "message", "interrupt"))) {
+      # if logged.conditions is NULL (default), do not log conditions
+      # if logged.conditions is a vector of strings, log only conditions which have their class in the vector
+      # if logged.conditions is NA, log all conditions
+      if (is.null(logged.conditions)
+          || (is.character(logged.conditions) && !inherits(c, logged.conditions)))
+        return()  # HACK (return not at end of function) to skip the following logging code
+    }
+
+
+
     log.message    <- c$message            # TODO: Should we use conditionMessage instead?
 
     if (is.null(log.message)) {
@@ -218,13 +240,22 @@ tryCatchLog <- function(expr,
     call.stack     <- sys.calls()          # "sys.calls" within "withCallingHandlers" is like a traceback!
     dump.file.name <- ""
 
-    # stack.trace <<- call.stack     # helper code for updating the expected result of the "test_build_log_entry" unit test
 
+
+    # stack.trace <<- call.stack     # helper code for updating the expected result of the "test_build_log_entry" unit test
     severity <-       if (inherits(c, "error"))     "ERROR"
                  else if (inherits(c, "warning"))   "WARN"
                  else if (inherits(c, "message"))   "INFO"
                  else if (inherits(c, "interrupt")) "INFO"
-                 else if (inherits(c, "condition")) "INFO"
+                 else if (inherits(c, "condition")) "INFO"  # TODO I would introduce a new logging level "DEBUG" or "VERBOSE" for this!
+                 #   # if logged.conditions is NULL (default), do not log conditions
+                 #   if (is.null(logged.conditions))     return()
+                 #   # if logged.conditions is NA, log all conditions
+                 #   else if ((length(logged.conditions) == 1) && is.na(logged.conditions))   "INFO"
+                 #   # if logged.conditions is a vector of strings, log only conditions which have their class in the vector
+                 #   else if (inherits(c, logged.conditions))   "INFO"
+                 #   else if (!inherits(c, logged.conditions))   return()
+                 # }
                  else stop(sprintf("Unsupported condition class %s!", class(c)))